# openapi.yml
openapi: 3.0.2
info:
  title: Сервис пользовательских рецептов
  description: Сервис пользовательских рецептов
  version: 0.1.0
paths:
...
  /recipe/:
    get:
      tags:
        - Recipes
      summary: Get Recipe
      operationId: get_recipe_recipe__get
      parameters:
        - required: false
          schema:
            title: Name
            type: string
          name: name
          in: query
        - required: false
          schema:
            title: Type
            type: string
          name: type
          in: query
        - required: false
          schema:
            title: Author
            type: integer
          name: author
          in: query
        - required: false
          schema:
            title: Page
            minimum: 1
            type: integer
            default: 1
          name: page
          in: query
        - required: false
          schema:
            title: Size
            maximum: 100
            minimum: 1
            type: integer
            default: 50
          name: size
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Union_RecipeListDB__NoneType__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Recipes
      summary: Create Recipe
      operationId: create_recipe_recipe__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDB'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OAuth2PasswordBearer: []
  /recipe/{id}:
    get:
      tags:
        - Recipes
      summary: Get Recipe
      operationId: get_recipe_recipe__id__get
      parameters:
        - required: true
          schema:
            title: Id
            type: integer
          name: id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDB'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    Page_Union_RecipeListDB__NoneType__:
      title: Page[Union[RecipeListDB, NoneType]]
      required:
        - items
        - total
        - page
        - size
      type: object
      properties:
        items:
          title: Items
          type: array
          items:
            $ref: '#/components/schemas/RecipeListDB'
        total:
          title: Total
          minimum: 0
          type: integer
        page:
          title: Page
          minimum: 1
          type: integer
        size:
          title: Size
          minimum: 1
          type: integer
...
    RecipeCreate:
      title: RecipeCreate
      required:
        - name
        - type
        - description
        - photo
        - steps
      type: object
      properties:
        name:
          title: Name
          type: string
        type:
          title: Type
          type: string
        description:
          title: Description
          type: string
        photo:
          title: Photo
          maxLength: 2083
          minLength: 1
          type: string
          format: uri
        steps:
          title: Steps
          type: string
    RecipeDB:
      title: RecipeDB
      required:
        - name
        - type
        - description
        - photo
        - id
        - user_id
        - create
        - update
        - status
        - steps
      type: object
      properties:
        name:
          title: Name
          type: string
        type:
          title: Type
          type: string
        description:
          title: Description
          type: string
        photo:
          title: Photo
          maxLength: 2083
          minLength: 1
          type: string
          format: uri
        id:
          title: Id
          type: integer
        user_id:
          title: User Id
          type: integer
        create:
          title: Create
          type: string
          format: date
        update:
          title: Update
          type: string
          format: date
        status:
          title: Status
          type: boolean
        steps:
          title: Steps
          type: string
    RecipeListDB:
      title: RecipeListDB
      required:
        - name
        - type
        - description
        - photo
        - id
        - user_id
        - create
        - update
        - status
      type: object
      properties:
        name:
          title: Name
          type: string
        type:
          title: Type
          type: string
        description:
          title: Description
          type: string
        photo:
          title: Photo
          maxLength: 2083
          minLength: 1
          type: string
          format: uri
        id:
          title: Id
          type: integer
        user_id:
          title: User Id
          type: integer
        create:
          title: Create
          type: string
          format: date
        update:
          title: Update
          type: string
          format: date
        status:
          title: Status
          type: boolean
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login